add custom exception

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -18,6 +19,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
         // verify that record is exists
+        if (leaveTypeToDelete == null)
+        {
+            throw new NotFoundException(nameof(leaveTypeToDelete), request.Id);
+        }
 
         // remove from database
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
